lca_analysis/pyproject.toml
[project]
name = "market-structure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Matthew Reda", email = "m.reda94@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "numba>=0.63.1",
    "numpyro>=0.19.0",
    "nutpie>=0.16.4",
    "plotly>=6.5.0",
    "prince>=0.16.2",
    "pymc>=5.26.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
streamlit = [
    "streamlit>=1.52.2",
]
dev = [
    "market-structure[all,streamlit]",
]

[tool.uv.sources]
market-structure = { workspace = true }
